modify imports after move file
